requirements.txt
streamlit==1.41.1
faiss-cpu==1.9.0.post1
numpy==2.2.2
sentence-transformers==3.4.0
PyPDF2==3.0.1
regex==2024.11.6
transformers==4.48.1